snabbkaffe (snabbkaffe v0.17.1+build.143.ref8a92a66)

This is the main API module

Types

event/0

Specs

event() :: #{'$kind' := kind(), '~meta' := map(), _ => _}.

filter/0

Specs

filter() :: predicate() | {predicate(), non_neg_integer()}.

kind/0

Specs

kind() :: atom() | string().

maybe/1

Specs

maybe(A) :: {just, A} | nothing.

maybe_pair/0

Specs

maybe_pair() ::
    {pair, timed_event(), timed_event()} | {unmatched_cause | unmatched_effect, timed_event()}.

metric/0

Specs

metric() :: atom().

predicate/0

Specs

predicate() :: fun((event()) -> boolean()).

predicate2/0

Specs

predicate2() :: fun((event(), event()) -> boolean()).

run_config/0

Specs

run_config() :: #{bucket => integer(), timeout => integer(), timetrap => integer()}.

timed_event/0

Specs

timed_event() :: #{'$kind' := kind(), ts := timestamp(), _ => _}.

timestamp/0

Specs

timestamp() :: integer().

trace/0

Specs

trace() :: [event()].

trace_spec/1

Specs

trace_spec(Result) :: fun((Result, trace()) -> ok | true) | fun((trace()) -> ok | true).

trace_specs/1

Specs

trace_specs(Result) :: trace_spec(Result) | [trace_spec(Result) | {string(), trace_spec(Result)}].

Functions

analyze_statistics()

Print metrics to the console

block_until(Filter, Timeout)

Specs

block_until(filter(), timeout()) -> {ok, event()} | timeout.

Equivalent to block_until(Filter, Timeout, infinity).

block_until(Filter, Timeout, BackInTime)

Specs

block_until(filter(), timeout(), timeout()) -> event() | timeout.

Block execution of the run stage of a testcase until an event matching Predicate is received or until Timeout.

Note: since the most common use case for this function is the following:

     trigger_produce_event_async(),
     snabbkaffe:block_until(MatchEvent, 1000)

there is a possible situation when the event is emitted before block_until function has a chance to run. In this case the latter will time out for no good reason. In order to work around this, block_until function actually searches for events matching Predicate in the past. BackInTime parameter determines how far back into past this function peeks.
Note: In the current implementation Predicate runs for every received event. It means this function should be lightweight

causality(Strict, CauseP, EffectP, Guard, Trace)

Specs

causality(boolean(),
          fun((event()) -> ok),
          fun((event()) -> ok),
          fun((event(), event()) -> boolean()),
          trace()) ->
             boolean().

check_conseq_relation(ErrorMsg, Relation, L)

Specs

check_conseq_relation(string(), fun((A, A) -> boolean()), [A]) -> boolean().

Throws an exception when consequitive elements of the list do not comply a binary relationship. Otherwise, returns true if the list is non-empty, and false when it is empty.
Example:
  SeqNums = ?projection(sequence_number, ?of_kind(handle_message, Trace)),
  ?assert(snabbkaffe:check_conseq_relation(SeqNums)),

cleanup()

Specs

cleanup() -> ok.

Cleanup the trace and the injected crashes and schedulings

collect_trace()

Specs

collect_trace() -> trace().

collect_trace(Timeout)

Specs

collect_trace(integer()) -> trace().

dump_trace(Trace)

Specs

dump_trace(trace()) -> file:filename().

Dump trace to a file and return the file name

erase_timestamps(Trace)

Specs

erase_timestamps(trace()) -> trace().

Remove timestamps from the trace events

events_of_kind(Kind, Events)

Specs

events_of_kind(kind() | [kind()], trace()) -> trace().

Extract events of certain kind(s) from the trace

find_pairs(CauseP, EffectP, Guard, L)

Specs

find_pairs(fun((event()) -> boolean()),
           fun((event()) -> boolean()),
           fun((event(), event()) -> boolean()),
           trace()) ->
              [maybe_pair()].

Find pairs of complimentary events

fix_ct_logging()

Specs

fix_ct_logging() -> ok.

Redirect all the logs to the console

forward_trace(Node)

Specs

forward_trace(node()) -> ok.

Forward traces from the remote node to the local node.

get_stats()

Specs

get_stats() -> #{snabbkaffe:metric() => snabbkaffe_collector:datapoints()}.

increasing(L)

Specs

increasing(list()) -> boolean().

Throws an exception when elements of the list are not increasing. Otherwise, returns true if the list is non-empty, and false when it is empty.
Example:
  SeqNums = ?projection(sequence_number, ?of_kind(handle_message, Trace)),
  ?assert(snabbkaffe:increasing(SeqNums)),

local_tp(Location, Level, Kind, Data, Metadata)

Specs

local_tp(term(), logger:level(), kind(), map(), map()) -> ok.

mk_all(Module)

Specs

mk_all(module()) -> [atom() | {group, atom()}].

Implement all/0 callback for Common Test

pair_max_depth(Pairs)

Specs

pair_max_depth([maybe_pair()]) -> non_neg_integer().

projection(Field, Trace)

Specs

projection([atom()] | atom(), trace()) -> list().

Extract specified field(s) from the list of events

projection_complete(Fields, Trace, Expected)

Specs

projection_complete(atom() | [atom()], trace(), [term()]) -> true.

projection_is_subset(Fields, Trace, Expected)

Specs

projection_is_subset(atom() | [atom()], trace(), [term()]) -> true.

push_stat(Metric, Num)

Specs

push_stat(metric(), number()) -> ok.

Report a scalar metric

push_stat(Metric, X, Y)

Specs

push_stat(metric(), number() | undefined, number()) -> ok.

Report a metric with an X value

push_stats(Metric, Pairs)

Specs

push_stats(metric(), [maybe_pair()] | number()) -> ok.

Report metrics from pairs of events

push_stats(Metric, Bucket, Pairs)

Specs

push_stats(metric(), number(), [maybe_pair()] | number()) -> ok.

Report multiple values

receive_events(SubRef)

Specs

receive_events(snabbkaffe_collector:subscription()) -> {ok | timeout, [snabbkaffe:event()]}.

Receive the events for a subscription SubRef

remote_tp(Location, Level, Kind, Data, Meta)

Specs

remote_tp(term(), logger:level(), kind(), map(), map()) -> ok.

retry(Interval, N, Fun)

Specs

retry(integer(), non_neg_integer(), fun(() -> Ret)) -> Ret.

Retry an action until success, at max N times with an interval Interval

splitl(Pred, L)

Specs

splitl(fun((A) -> boolean()), [A]) -> [[A]].

Split list by predicate like this:
     [true, true, false, true, true, false] ->
      [[true, true, false], [true, true, false]]

splitr(Pred, L)

Specs

splitr(fun((A) -> boolean()), [A]) -> [[A]].

Split list by predicate like this:
     [true, true, false, true, true, false] ->
      [[true, true], [false, true, true], [false]]

start_trace()

Specs

start_trace() -> ok.

stop()

Specs

stop() -> ok.

Stop snabbkaffe

strictly_increasing(L)

Specs

strictly_increasing(list()) -> boolean().

Throws an exception when elements of the list are not strictly increasing. Otherwise, returns true if the list is non-empty, and false when it is empty.
Example:
  SeqNums = ?projection(sequence_number, ?of_kind(handle_message, Trace)),
  ?assert(snabbkaffe:strictly_increasing(SeqNums)),

subscribe(Predicate)

Specs

subscribe(snabbkaffe:predicate()) -> {ok, snabbkaffe_collector:subscription()}.

Equivalent to subscribe(Predicate, 1, infinity, 0).

subscribe(Predicate, Timeout)

Specs

subscribe(snabbkaffe:predicate(), timeout()) -> {ok, snabbkaffe_collector:subscription()}.

Equivalent to subscribe(Predicate, 1, Timeout, 0).

subscribe(Predicate, NEvents, Timeout)

Specs

subscribe(snabbkaffe:predicate(), non_neg_integer(), timeout()) ->
             {ok, snabbkaffe_collector:subscription()}.

Equivalent to subscribe(Predicate, NEvents, Timeout, 0).

subscribe(Predicate, NEvents, Timeout, BackInTime)

Specs

subscribe(snabbkaffe:predicate(), non_neg_integer(), timeout(), timeout()) ->
             {ok, snabbkaffe_collector:subscription()}.

Subscribe to events matching Predicate`, receive `NEvents of this type, or until Timeout happens, look at most BackInTime milliseconds into the past.
The events can be received using receive_events/1 function.

tp(Location, Level, Kind, Data)

Specs

tp(term(), logger:level(), kind(), map()) -> ok.

unique(Trace)

Specs

unique(trace()) -> true.

Throw an exception if some event is repeated

wait_async_action(Action, Predicate, Timeout)

Specs

wait_async_action(fun(() -> Return), predicate(), timeout()) -> {Return, {ok, event()} | timeout}.

Install a listener for the event matching Predicate, execute an action, then wait for the event to arrive, or until Timeout.