snabbkaffe_nemesis (snabbkaffe v0.17.1+build.143.ref8a92a66)

This module implements "nemesis" process that injects faults into system under test in order to test its fault-tolerance.

Usage

Somewhere in the tested code

  ?maybe_crash(kind1, #{data1 => Foo, field2 => Bar})

Somewhere in the run stage

  ?inject_crash( #{?snk_kind := kind1, data1 := 42}
               , snabbkaffe_nemesis:always_crash()
               )

Types

fault_key/0

Specs

fault_key() :: term().

fault_scenario/0
(opaque)

Specs

fault_scenario()

Functions

always_crash()

Specs

always_crash() -> fault_scenario().

cleanup()

Specs

cleanup() -> ok.

Remove all injected crashes

fix_crash(Ref)

Specs

fix_crash(reference()) -> ok.

Remove injected fault

force_ordering(DelayPredicate, NEvents, ContinuePredicate)

Specs

force_ordering(snabbkaffe:predicate(), non_neg_integer(), snabbkaffe:predicate2()) -> reference().

Inject delay into the system

inject_crash(Predicate, Scenario)

Specs

inject_crash(snabbkaffe:predicate(), fault_scenario()) -> reference().

Equivalent to inject_crash(Predicate, Scenario, notmyday).

inject_crash(Predicate, Scenario, Reason)

Specs

inject_crash(snabbkaffe:predicate(), fault_scenario(), term()) -> reference().

Inject crash into the system

maybe_crash(Key, Data)

Specs

maybe_crash(fault_key(), map()) -> ok.

Check if there are any injected crashes that match this data, and respond with the crash reason if so.

maybe_delay(Event)

Specs

maybe_delay(map()) -> ok.

Check if the trace point should be delayed.

periodic_crash(Period, DutyCycle, Phase)

Specs

periodic_crash(integer(), float(), float()) -> fault_scenario().

A type of fault that occurs and fixes periodically.

random_crash(CrashProbability)

Specs

random_crash(float()) -> fault_scenario().

recover_after(Times)

Specs

recover_after(non_neg_integer()) -> fault_scenario().

start_link()

Start the server