snabbkaffe_nemesis (snabbkaffe v1.0.7)

This module implements "nemesis" process that injects faults into system under test in order to test its fault-tolerance.

Usage

Somewhere in the tested code

  ?maybe_crash(kind1, #{data1 => Foo, field2 => Bar})

Somewhere in the run stage

  ?inject_crash( #{?snk_kind := kind1, data1 := 42}
               , snabbkaffe_nemesis:always_crash()
               )

Types

fault_key/0

-type fault_key() :: term().

fault_scenario/0
(opaque)

-opaque fault_scenario()

Functions

always_crash()

-spec always_crash() -> fault_scenario().

cleanup()

-spec cleanup() -> ok.

Remove all injected crashes

fix_crash(Ref)

-spec fix_crash(reference()) -> ok.

Remove injected fault

force_ordering(DelayPredicate, NEvents, ContinuePredicate)

-spec force_ordering(snabbkaffe:predicate(), non_neg_integer(), snabbkaffe:predicate2()) -> reference().

Inject scheduling into the system

inject_crash(Predicate, Scenario)

-spec inject_crash(snabbkaffe:predicate(), fault_scenario()) -> reference().

Equivalent to inject_crash(Predicate, Scenario, notmyday).

inject_crash(Predicate, Scenario, Reason)

-spec inject_crash(snabbkaffe:predicate(), fault_scenario(), term()) -> reference().

Inject crash into the system

periodic_crash(Period, DutyCycle, Phase)

-spec periodic_crash(integer(), float(), float()) -> fault_scenario().

A type of fault that occurs and fixes periodically.

random_crash(CrashProbability)

-spec random_crash(float()) -> fault_scenario().

Crash with probability P

recover_after(Times)

-spec recover_after(non_neg_integer()) -> fault_scenario().

Crash N times consequently, then recover